First working versión with local isntallation!!!
This, by creating an aditional install() function. The code here is repeated and needs to be reduced, maybe
in only one function for the local and systemwide installation

Also added descriptibe info if the user tries to install without sudo.

A TODO list added at the begining of the program.